mcp_server/requirements.txt
fastmcp>=0.1.0
homeassistant>=2024.1.0
uvicorn>=0.24.0
httpx>=0.25.0

